Use separate files for argument classes
